LPS-77701 portal-search: Merge SearchResultPermissionFiler files, since there is currently only one class
